Reduce building of PureScript module map
Resolves #59 and resolves #60
