+++ .gitignore
@@ -1,3 +1,3 @@
 **DS_Store*
 node_modules/
-index.js
+lib/
